app/src/requirements.txt
# This file is used to install the required packages for the app
flask==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
